refactor: Car 클래스 및 테스트 클래스 리팩터링
 - Car 클래스 메서드명 변경 및 문자열을 Name 클래스로 변경
 - CarsTest 클래스에서 DisplayName 추가 및 메서드 이름 변경
 - CarTest에서 Car 클래스 메서드명 변경에 따른 리팩터링
